+++ src/main/kotlin/com/example/ramenbm/global/security/filter/JwtRequestFilter.kt
@@ -1,6 +1,6 @@
 package com.example.ramenbm.global.security.filter
 
-import com.example.ramenbm.global.security.jwt.TokenProvider
+import com.example.ramenbm.global.security.jwt.TokenParser
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.stereotype.Component
 import org.springframework.web.filter.OncePerRequestFilter
@@ -10,7 +10,7 @@ import javax.servlet.http.HttpServletResponse
 
 @Component
 class JwtRequestFilter(
-        private val tokenProvider: TokenProvider
+        private val tokenParser: TokenParser
 ): OncePerRequestFilter() {
 
     override fun doFilterInternal(
@@ -18,9 +18,9 @@ class JwtRequestFilter(
             response: HttpServletResponse,
             filterChain: FilterChain
     ) {
-        val accessToken = tokenProvider.parseAccessToken(request)
+        val accessToken = tokenParser.parseAccessToken(request)
         if (!accessToken.isNullOrBlank()) {
-            val authentication = tokenProvider.authentication(accessToken)
+            val authentication = tokenParser.authentication(accessToken)
             val securityContext = SecurityContextHolder.getContext()
             securityContext.authentication = authentication
         }
